MINTEMP/MAXTEMP: fix code to skip if PS_ON
Current code checks to see if there's a PS_ON pin so it can
skip the MINTEMP and MAXTEMP checks.  This is because the
PS_ON feature on the Gen7 causes bogus temperatures to be
read when the PS is OFF.

The logic is backwards, though.  Fix the logic so these
checks are executed when we DO NOT have a PS_ON pin rather
than when we DO.